moves events.go files to v1/common pkg
