+++ api/pkg/handler/handler_test.go
@@ -1,4 +1,4 @@
-package handler_test
+package handler
 
 import (
 	"context"
@@ -6,8 +6,6 @@ import (
 	"net/http/httptest"
 	"strings"
 	"testing"
-
-	"github.com/kubermatic/kubermatic/api/pkg/handler"
 )
 
 func TestEncodeJSON(t *testing.T) {
@@ -29,7 +27,7 @@ func TestEncodeJSON(t *testing.T) {
 	for _, testcase := range testcases {
 		writer := httptest.NewRecorder()
 
-		err := handler.EncodeJSON(ctx, writer, testcase.input)
+		err := encodeJSON(ctx, writer, testcase.input)
 		if err != nil {
 			t.Errorf("failed to encode %#v as JSON: %v", testcase.input, err)
 		}
